requirements.txt
xarray
matplotlib
cartopy
cfgrib
ecmwf-opendata
numpy
geopandas
